fix: 주거 옵션 요청의 API 타입과 퍼널 도메인 타입 분리
