+++ src/pages/imageSetup/types/apis/houseInfo.ts
@@ -10,13 +10,6 @@ export interface HousingOptionsResponse {
   areaTypes: HousingOptionItem[];
 }
 
-// 유니온 타입 추출
-export type HouseTypeCode =
-  HousingOptionsResponse['houseTypes'][number]['code'];
-// "OFFICETEL" | "VILLA" | "APARTMENT" | "ETC" 유니온 타입
-export type RoomTypeCode = HousingOptionsResponse['roomTypes'][number]['code'];
-export type AreaTypeCode = HousingOptionsResponse['areaTypes'][number]['code'];
-
 // 주거 선택 전송 (POST)
 export interface HousingSelectionRequest extends Record<string, unknown> {
   houseType: string;
